requirements.txt
django

# Don't forget to add "taggit" to your Installed apps.
django-taggit

